chore: deduplicate client/server context helpers
