+++ .changeset/shared-context-helpers.md
@@ -1,0 +1,5 @@
+---
+'svelte': patch
+---
+
+chore: deduplicate client and server context helpers
